First attempt at generic buffer
This is the first attempt to move from a plain old `u8` buffer to a
generic type buffer. It has an uninitialized memory free bug when using
non-`Copy`(i think this is the right trigger) types because the vector
used for the buffer doesnt know we spoofed its length and tries to `drop`
the value we would be "overwriting" if it existed.