Busy waiting instead of signaled waiting
This replaces the signal based for reading an empty or writing a full
buffer with a busy waiting scheme where the waiter keeps yeilding to
keep actual wasted time down for slow readers or slow writers but greatly
reducing contention for fast writers and readers which will be common case.
This version runs in equivalent load and time as mutex waiting for medium sized writers and readers.